+++ database/upgrade/upgrade4.10.0.sql
@@ -9,7 +9,14 @@ START TRANSACTION;
 ALTER TABLE myems_system_db.tbl_spaces
 ADD `number_of_occupants` DECIMAL(18, 3) NOT NULL DEFAULT 1.000 AFTER `area`;
 
+CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_spaces_distribution_systems` (
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
+  `space_id` BIGINT NOT NULL,
+  `distribution_system_id` BIGINT NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_spaces_distribution_systems_index_1` ON `myems_system_db`.`tbl_spaces_distribution_systems` (`space_id`);
+
 -- UPDATE VERSION NUMBER
-UPDATE `myems_system_db`.`tbl_versions` SET version='4.10.0RC', release_date='2024-10-30' WHERE id=1;
+UPDATE `myems_system_db`.`tbl_versions` SET version='4.10.0', release_date='2024-10-26' WHERE id=1;
 
 COMMIT;