+++ myems-web/src/components/MyEMS/dashboard/Dashboard.js
@@ -733,7 +733,7 @@ const Dashboard = ({ setRedirect, setRedirectUrl, t }) => {
       <div className="card-deck">
         {settings.showOnlineMap ? (
           <div className="mb-3 card" style={{ height: '400px' }}>
-            <CustomizeMapBox Latitude={rootLatitude} Longitude={rootLongitude} Zoom={3} Geojson={geojson['features']} />
+            <CustomizeMapBox Latitude={rootLatitude} Longitude={rootLongitude} Zoom={10} Geojson={geojson['features']} />
           </div>
         ) : (
           <></>